requirements.txt
streamlit
pandas
seaborn
psycopg-2binary
python-dotenv